Remove the old StudentCourses panel. Use the new panels and style. [#5858]
