--- lib/ui/academics/StudentCourses.dart
@@ -1,641 +1,0 @@
-import 'package:collection/collection.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-import 'package:illinois/ext/Canvas.dart';
-import 'package:illinois/ext/StudentCourse.dart';
-import 'package:illinois/ext/Explore.dart';
-import 'package:illinois/model/Analytics.dart';
-import 'package:illinois/model/Canvas.dart';
-import 'package:illinois/model/StudentCourse.dart';
-import 'package:illinois/service/Analytics.dart';
-import 'package:illinois/service/Auth2.dart';
-import 'package:illinois/service/Canvas.dart';
-import 'package:illinois/service/Config.dart';
-import 'package:illinois/service/Map2.dart';
-import 'package:illinois/service/StudentCourses.dart';
-import 'package:illinois/ui/canvas/CanvasCourseAssignmentsPanel.dart';
-import 'package:illinois/ui/explore/DisplayFloorPlanPanel.dart';
-import 'package:illinois/ui/home/HomeWidgets.dart';
-import 'package:illinois/ui/map2/Map2HomePanel.dart';
-import 'package:illinois/ui/map2/Map2Widgets.dart';
-import 'package:illinois/ui/widgets/AccentCard.dart';
-import 'package:illinois/ui/widgets/HeaderBar.dart';
-import 'package:illinois/ui/widgets/TabBar.dart' as uiuc;
-import 'package:illinois/utils/AppUtils.dart';
-import 'package:illinois/utils/Utils.dart';
-import 'package:rokwire_plugin/service/connectivity.dart';
-import 'package:rokwire_plugin/service/localization.dart';
-import 'package:rokwire_plugin/service/notification_service.dart';
-import 'package:rokwire_plugin/service/styles.dart';
-import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
-import 'package:rokwire_plugin/utils/utils.dart';
-import 'package:sprintf/sprintf.dart';
-
-class StudentCoursesContentWidget extends StatefulWidget with AnalyticsInfo {
-  StudentCoursesContentWidget();
-
-  @override
-  State<StudentCoursesContentWidget> createState() => _StudentCoursesContentWidgetState();
-
-  @override
-  AnalyticsFeature? get analyticsFeature => AnalyticsFeature.AcademicsStudentCourses;
-}
-
-class _StudentCoursesContentWidgetState extends State<StudentCoursesContentWidget> with NotificationsListener {
-
-  List<StudentCourse>? _courses;
-  bool _loading = false;
-
-  @override
-  void initState() {
-    NotificationService().subscribe(this, [
-      Auth2.notifyLoginChanged,
-      Connectivity.notifyStatusChanged,
-      StudentCourses.notifyTermsChanged,
-      StudentCourses.notifySelectedTermChanged,
-      StudentCourses.notifyCachedCoursesChanged,
-    ]);
-
-    if (Connectivity().isNotOffline && (StudentCourses().displayTermId != null) && Auth2().isOidcLoggedIn) {
-      _loading = true;
-      StudentCourses().loadCourses(termId: StudentCourses().displayTermId!, forceLoad: true).then((List<StudentCourse>? courses) {
-        setStateIfMounted(() {
-          _courses = courses;
-          _sortCourses();
-          _loading = false;
-        });
-      });
-    }
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    NotificationService().unsubscribe(this);
-    super.dispose();
-  }
-
-  // NotificationsListener
-
-  @override
-  void onNotification(String name, dynamic param) {
-    if (name == Auth2.notifyLoginChanged) {
-      _updateCourses();
-    }
-    else if (name == Connectivity.notifyStatusChanged) {
-      _updateCourses();
-    }
-    else if (name == StudentCourses.notifyTermsChanged) {
-      setStateIfMounted(() {});
-    }
-    else if (name == StudentCourses.notifySelectedTermChanged) {
-      _updateCourses();
-    }
-    else if (name == StudentCourses.notifyCachedCoursesChanged) {
-      if ((param == null) || (StudentCourses().displayTermId == param)) {
-        _updateCourses(forceLoad: false);
-      }
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return _buildContent();
-  }
-
-  Widget _buildContent() {
-    if (_loading) {
-      return _buildLoadingContent();
-    }
-    else if (Connectivity().isOffline) {
-      return _buildMessageContent(Localization().getStringEx('panel.student_courses.load.offline.error.msg', 'My Courses not available while offline.'));
-    }
-    else if (!Auth2().isOidcLoggedIn) {
-      return _buildMessageContent(AppTextUtils.loggedOutFeatureNA(Localization().getStringEx('generic.app.feature.student_courses', 'My Courses'), verbose: true));
-    }
-    else if (_courses == null) {
-      return _buildMessageContent(Localization().getStringEx('panel.student_courses.load.failed.error.msg', 'It appears you have no courses registered for the selected term.'));
-    }
-    else if (_courses?.isEmpty ?? true) {
-      return _buildMessageContent(Localization().getStringEx('panel.student_courses.empty.content.msg', 'You do not appear to be enrolled in any courses for the selected term.'));
-    }
-    else {
-      return _buildCoursesContent();
-    }
-  }
-
-  Widget _buildCoursesContent() {
-    List<Widget> courseWidgets = <Widget>[
-      Align(alignment: Alignment.centerLeft, child:
-        _buildTermsDropDown(),
-      ),
-    ];
-    
-    if (_courses != null) {
-      for (StudentCourse course in _courses!) {
-        courseWidgets.add(Padding(padding: EdgeInsets.only(top: (1 < courseWidgets.length) ? 8 : 0), child:
-          StudentCourseCard(course: course, analyticsFeature: widget.analyticsFeature,),
-        ));
-
-      }
-    }
-
-    return SingleChildScrollView(child:
-      Padding(padding: EdgeInsets.only(bottom: 16), child:
-        Column(children: courseWidgets,),
-      )
-    );
-  }
-
-  Widget _buildLoadingContent() {
-    return Column(children: <Widget>[
-      Align(alignment: Alignment.centerLeft, child:
-        _buildTermsDropDown(),
-      ),
-      Expanded(flex: 1, child: Container()),
-      CircularProgressIndicator(strokeWidth: 3, valueColor: AlwaysStoppedAnimation<Color?>(Styles().colors.fillColorSecondary),),
-      Expanded(flex: 4, child: Container()),
-    ]);
-  }
-
-  Widget _buildMessageContent(String message) {
-    return Column(children: <Widget>[
-      Align(alignment: Alignment.centerLeft, child:
-        _buildTermsDropDown(),
-      ),
-      Expanded(flex: 1, child: Container()),
-      Padding(padding: EdgeInsets.symmetric(horizontal: 28), child:
-        Center(child:
-          Text(message, textAlign: TextAlign.center, style: Styles().textStyles.getTextStyle("widget.message.medium.thin"))
-        ),
-      ),
-      Expanded(flex: 4, child: Container()),
-    ]);
-  }
-
-  TextStyle? getTermDropDownItemStyle({bool selected = false}) => selected ?
-  Styles().textStyles.getTextStyle("widget.message.regular") :
-  Styles().textStyles.getTextStyle("widget.message.regular.semi_fat");
-
-  Widget _buildTermsDropDown() {
-    StudentCourseTerm? currentTerm = StudentCourses().displayTerm;
-
-    return Visibility(visible: CollectionUtils.isNotEmpty(StudentCourses().terms), child:
-      Semantics(label: currentTerm?.name, hint: "Double tap to select account", button: true, container: true, child:
-        DropdownButtonHideUnderline(child:
-          DropdownButton<String>(
-            icon: Padding(padding: EdgeInsets.only(left: 4), child: Styles().images.getImage('chevron-down', excludeFromSemantics: true)),
-            isExpanded: false,
-            style: getTermDropDownItemStyle(selected: true),
-            //alignment: AlignmentDirectional.centerEnd,
-            hint: (currentTerm?.name?.isNotEmpty ?? false) ? Text(currentTerm?.name ?? '', style: getTermDropDownItemStyle(selected: true)) : null,
-            items: _buildTermDropDownItems(),
-            onChanged: _onTermDropDownValueChanged
-          ),
-        ),
-      ),
-    );
-  }
-
-  List<DropdownMenuItem<String>>? _buildTermDropDownItems() {
-    List<StudentCourseTerm>? terms = StudentCourses().terms;
-    String? currentTermId = StudentCourses().displayTermId;
-
-    List<DropdownMenuItem<String>>? items;
-    if (terms != null) {
-      items = <DropdownMenuItem<String>>[];
-      for (StudentCourseTerm term in terms) {
-        items.add(DropdownMenuItem<String>(
-          value: term.id,
-          child: Text(term.name ?? '', style: getTermDropDownItemStyle(selected: term.id == currentTermId),)
-        ));
-      }
-    }
-    return items;
-  }
-
-  void _onTermDropDownValueChanged(String? termId) {
-    StudentCourses().selectedTermId = termId;
-  }
-
-  void _updateCourses({bool forceLoad = true}) {
-    if (Connectivity().isNotOffline && (StudentCourses().displayTermId != null) && Auth2().isOidcLoggedIn) {
-      setStateIfMounted(() {
-        _loading = true;
-      });
-      StudentCourses().loadCourses(termId: StudentCourses().displayTermId!, forceLoad: forceLoad).then((List<StudentCourse>? courses) {
-        setStateIfMounted(() {
-          _courses = courses;
-          _sortCourses();
-          _loading = false;
-        });
-      });
-    }
-    else {
-        setStateIfMounted(() {});
-    }
-  }
-
-  void _sortCourses() {
-    _courses?.sort((StudentCourse c1, StudentCourse c2) {
-      String n1 = c1.section?.comparableValue ?? '';
-      String n2 = c2.section?.comparableValue ?? '';
-      return n1.compareTo(n2);
-    });
-  }
-}
-
-class StudentCourseCard extends StatelessWidget {
-  final StudentCourse course;
-  final CardDisplayMode displayMode;
-  final AnalyticsFeature? analyticsFeature;
-  
-  StudentCourseCard({super.key,
-    required this.course,
-    this.displayMode = CardDisplayMode.browse,
-    this.analyticsFeature
-  });
-
-  static double height(BuildContext context) =>
-    MediaQuery.of(context).textScaler.scale(36 + 18 + (6 + 16) + 16 + (6 + 18) + (12 + 18));
-
-  @override
-  Widget build(BuildContext context) =>
-    InkWell(onTap: () => _onCard(context), child:
-      Semantics(label: course.title,
-        child: AccentCard(
-          displayMode: displayMode,
-          accentColor: Styles().colors.fillColorSecondary,
-          child: _contentWidget,
-        )
-      ),
-    );
-
-  Widget get _contentWidget {
-    String courseSchedule = course.section?.displaySchedule ?? '';
-    String courseLocation = course.section?.displayLocation ?? '';
-
-    return Padding(padding: EdgeInsets.all(16), child:
-      Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-
-        Row(children: [Expanded(child:
-          Text(course.title ?? '', style: Styles().textStyles.getTextStyle("widget.card.title.regular.extra_fat")),
-        )]),
-
-        Padding(padding: EdgeInsets.only(top: 6), child:
-          Row(children: [Expanded(child:
-            Text(course.displayInfo, style: Styles().textStyles.getTextStyle("widget.card.detail.medium")),
-          )]),
-        ),
-
-        Padding(padding: EdgeInsets.zero, child:
-          Row(children: [Expanded(child:
-            Text(sprintf(Localization().getStringEx('panel.student_courses.instructor.title', 'Instructor: %s'), [course.section?.instructor ?? '']), style: Styles().textStyles.getTextStyle("widget.card.detail.medium"),)
-          )]),
-        ),
-
-        Visibility(visible: courseSchedule.isNotEmpty, child:
-          Padding(padding: EdgeInsets.only(top: 6), child:
-            Row(children: [
-              Padding(padding: EdgeInsets.only(right: 6), child:
-                Styles().images.getImage('calendar', excludeFromSemantics: true),
-              ),
-              Expanded(child:
-                Text(courseSchedule, style: Styles().textStyles.getTextStyle("widget.card.detail.medium")),
-              )
-
-            ],),
-          ),
-        ),
-
-        Visibility(visible: courseLocation.isNotEmpty, child:
-          InkWell(onTap: course.hasValidLocation ? _onLocaltion : null, child:
-            Padding(padding: EdgeInsets.symmetric(vertical: 6), child:
-              Row(children: [
-                Padding(padding: EdgeInsets.only(right: 6), child:
-                  Styles().images.getImage('location', excludeFromSemantics: true),
-                ),
-                Expanded(child:
-                  Text(courseLocation, style: course.hasValidLocation ?
-                    Styles().textStyles.getTextStyle("widget.button.light.title.medium.underline") :
-                    Styles().textStyles.getTextStyle("widget.button.light.title.medium")
-                  ),
-                )
-              ],),
-            ),
-          ),
-        ),
-
-        Visibility(visible: (course.section?.isInPerson == true) && course.hasValidLocation, child:
-          Padding(padding: EdgeInsets.symmetric(vertical: 6), child:
-            Row(children: [
-              Padding(padding: EdgeInsets.only(right: 6), child:
-                Opacity(opacity: 0, child:
-                  Styles().images.getImage('location', excludeFromSemantics: true),
-                )
-              ),
-              Expanded(child:
-                Wrap(spacing: 8, runSpacing: 8, children: [
-                  Map2NavDirectionsButton('person-walking', onTap: () => _onNavigationDirections(GeoMapUtils.traveModeWalking)),
-                  Map2NavDirectionsButton('bicycle', onTap: () => _onNavigationDirections(GeoMapUtils.traveModeBycycling)),
-                  Map2NavDirectionsButton('car', onTap: () => _onNavigationDirections(GeoMapUtils.traveModeDriving)),
-                  Map2NavDirectionsButton('bus', onTap: () => _onNavigationDirections(GeoMapUtils.traveModeTransit)),
-                ],),
-              )
-            ],),
-          ),
-        ),
-      ],)
-    );
-  }
-
-  void _onLocaltion() {
-    Analytics().logSelect(target: "Location Detail");
-    course.launchDirections();
-  }
-
-  void _onNavigationDirections(String travelMode) {
-    Analytics().logSelect(target: 'Navigation Directions: $travelMode');
-    course.launchDirections(travelMode: travelMode);
-    //GeoMapUtils.launchDirections(destination: course?.section?.building?.exploreLocation?.exploreLocationMapCoordinate, travelMode: travelMode);
-  }
-
-  void _onCard(BuildContext context) {
-    Analytics().logSelect(target: "Student Course: ${course.title}");
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => StudentCourseDetailPanel(course: course, analyticsFeature: analyticsFeature,)));
-  }
-
-}
-
-class StudentCoursesListPanel extends StatelessWidget with AnalyticsInfo {
-  StudentCoursesListPanel();
-
-  @override
-  AnalyticsFeature? get analyticsFeature => AnalyticsFeature.AcademicsStudentCourses;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: HeaderBar(title: Localization().getStringEx('panel.student_courses.header.title', 'My Courses')),
-      body: Padding(padding: EdgeInsets.only(left: 16, right: 16, bottom: 16), child: StudentCoursesContentWidget()),
-      backgroundColor: Styles().colors.white,
-      bottomNavigationBar: uiuc.TabBar()
-    );
-  }
-}
-
-class StudentCourseDetailPanel extends StatefulWidget with AnalyticsInfo {
-  final StudentCourse? course;
-  final AnalyticsFeature? analyticsFeature; //This overrides AnalyticsInfo.analyticsFeature getter
-
-  StudentCourseDetailPanel({super.key, this.course, this.analyticsFeature});
-  @override
-  _StudentCourseDetailPanelState createState() => _StudentCourseDetailPanelState();
-}
-class _StudentCourseDetailPanelState extends State<StudentCourseDetailPanel> {
-  bool _floorsPlansExpanded = false;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: _buildContent(),
-      backgroundColor: Styles().colors.white,
-      bottomNavigationBar: uiuc.TabBar()
-    );
-  }
-
-  Widget _buildContent() {
-    return Column(children: <Widget>[
-      Expanded(child:
-        Container(child:
-          CustomScrollView(scrollDirection: Axis.vertical, slivers: <Widget>[
-            SliverToutHeaderBar(
-              flexRightToLeftTriangleColor: Colors.white,
-              flexImageKey: 'course-detail-default',
-            ),
-            SliverList(delegate: SliverChildListDelegate([
-              Container(color: Colors.white, padding:EdgeInsets.symmetric(horizontal: 20), child:
-                Column(crossAxisAlignment: CrossAxisAlignment.start, children: <Widget>[
-                  _buildTitle(),
-
-                  if ((widget.course?.displayInfo.isNotEmpty == true) || (widget.course?.displayInfo.isNotEmpty == true))
-                    ...<Widget>[
-                      Container(height: 12),
-                      if (widget.course?.displayInfo.isNotEmpty == true)
-                        _buildDisplayInfo(),
-                      if (widget.course?.section?.instructor?.isNotEmpty == true)
-                        _buildInstructor(),
-                    ],
-
-                  Container(height: 18),
-
-                  if (widget.course?.section?.displaySchedule.isNotEmpty == true)
-                    _buildScheduleDetail(),
-
-                  if ((widget.course?.section?.isInPerson == true) && (widget.course?.section?.building?.fullAddress?.isNotEmpty == true))
-                    _buildLocation(),
-
-                  if ((widget.course?.section?.isInPerson == true) && (widget.course?.section?.building?.floors?.isNotEmpty == true))
-                    _buildFloorPlans(),
-
-                  if ((widget.course?.section?.isInPerson == true) && (widget.course?.hasValidLocation == true))
-                    _buildNavDirections(),
-
-                  if (_canvasVisible)
-                    _buildCanvasContent(),
-
-                  Container(height: 32),
-                  
-                  if (widget.course?.section?.isInPerson == true)
-                    _buildInPersonMapButton(),
-                  
-                ])
-              )
-            ], addSemanticIndexes:false),),
-          ],),
-        ),
-      ),
-    ],);
-  }
-
-  Widget _buildTitle() =>
-    Text(widget.course?.title ?? "", style: Styles().textStyles.getTextStyle("widget.title.medium_large.fat"));
-
-  Widget _buildDisplayInfo() =>
-    Text(widget.course?.displayInfo ?? "", style: Styles().textStyles.getTextStyle("widget.item.regular.thin"));
-
-  Widget _buildInstructor() =>
-    Text(sprintf(Localization().getStringEx('panel.student_courses.instructor.title', 'Instructor: %s'), [widget.course?.section?.instructor ?? '']), style: Styles().textStyles.getTextStyle("widget.item.regular.thin"),);
-
-  Widget _buildScheduleDetail() =>
-    Padding(padding: EdgeInsets.symmetric(vertical: 6), child:
-      Row(children: [
-        _buildDetailIcon('calendar'),
-        Expanded(child:
-          Text(widget.course?.section?.displaySchedule ?? '', style: Styles().textStyles.getTextStyle("widget.item.regular.thin")),
-        )
-      ],),
-    );
-
-  Widget _buildLocation() =>
-    InkWell(onTap: (widget.course?.hasValidLocation ?? false) ? _onLocation : null, child:
-      Padding(padding: EdgeInsets.symmetric(vertical: 6,), child:
-        Row(children: [
-          _buildDetailIcon('location'),
-          Expanded(child:
-            Text(widget.course?.section?.building?.fullAddress ?? '', style: (widget.course?.hasValidLocation ?? false) ?
-              Styles().textStyles.getTextStyle("widget.button.light.title.medium.underline") :
-              Styles().textStyles.getTextStyle("widget.button.light.title.medium")
-            ),
-          )
-        ],),
-      ),
-    );
-
-
-  Widget _buildNavDirections() =>
-    Padding(padding: EdgeInsets.symmetric(vertical: 6,), child:
-      Row(children: [
-        _buildDetailIcon(),
-        Expanded(child:
-          Wrap(spacing: 8, runSpacing: 8, children: [
-            Map2NavDirectionsButton('person-walking', onTap: () => _onNavigationDirections(GeoMapUtils.traveModeWalking)),
-            Map2NavDirectionsButton('bicycle', onTap: () => _onNavigationDirections(GeoMapUtils.traveModeBycycling)),
-            Map2NavDirectionsButton('car', onTap: () => _onNavigationDirections(GeoMapUtils.traveModeDriving)),
-            Map2NavDirectionsButton('bus', onTap: () => _onNavigationDirections(GeoMapUtils.traveModeTransit)),
-          ],),
-        )
-      ],),
-    );
-
-  Widget _buildFloorPlans() {
-    String? room = widget.course?.section?.room;
-    String displayRoom = ((room != null) && room.isNotEmpty) ? room : Localization().getStringEx('panel.student_courses.room.title', 'Room');
-    List<String> floors = widget.course?.section?.building?.floors ?? [];
-    return Padding(padding: EdgeInsets.symmetric(), child:
-      Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-        InkWell(onTap: _onFloorsPlans, child:
-          Padding(padding: EdgeInsets.symmetric(vertical: 6), child:
-            Row(children: [
-              _buildDetailIcon(_floorsPlansExpanded ? 'chevron-up' : 'chevron-down'),
-              Text("Room ${displayRoom}", style: Styles().textStyles.getTextStyle("widget.button.light.title.medium.underline"))
-            ]),
-          ),
-        ),
-        Visibility(visible: _floorsPlansExpanded, child:
-          Row(children: [
-            _buildDetailIcon(),
-            Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-              Text("${Localization().getStringEx('panel.explore_building_detail.detail.fllor_plan_and_amenities', 'Floor Plans & Amenities')}:", style: Styles().textStyles.getTextStyle('widget.button.light.title.medium.fat')),
-              ...floors.map((floor) => InkWell(onTap: () => _onFloor(floor), child:
-                Padding(padding: EdgeInsets.symmetric(vertical: 2), child:
-                  Text("Floor ${floor}", style: Styles().textStyles.getTextStyle("widget.description.small.underline")))
-                )
-              ).toList()
-            ])
-          ]),
-        )
-      ])
-    );
-  }
-
-  Widget _buildDetailIcon([String? iconKey]) =>
-    Padding(padding: EdgeInsets.only(right: 8), child:
-      SizedBox(width: 18, height: 18, child:
-        Center(child: (iconKey != null) ?
-          Styles().images.getImage(iconKey, excludeFromSemantics: true) : null
-        )
-      )
-    );
-
-  Widget _buildInPersonMapButton() =>
-    Padding(padding: EdgeInsets.symmetric(horizontal: 18 + 8), child:
-      RoundedButton(
-        label: Localization().getStringEx('panel.student_courses.map.button.title', 'View In-Person Courses on Map'),
-        textStyle: Styles().textStyles.getTextStyle('widget.button.title.medium.fat'),
-        padding: EdgeInsets.symmetric(vertical: 8, horizontal: 16),
-        onTap: _onMap,
-      ),
-    );
-
-  Widget _buildCanvasContent() {
-    return Padding(padding: EdgeInsets.only(top: 16), child: Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
-      Container(height: 1, color: Styles().colors.surfaceAccent2,),
-      _buildCanvasButton(label: Localization().getStringEx('panel.student_courses.canvas.assignments.view.label', 'View Assignments in Canvas'), onTap: _onCanvasAssignments),
-      _buildCanvasButton(label: Localization().getStringEx('panel.student_courses.canvas.launch.view.label', 'Launch Canvas'), onTap: _onCanvasLaunch),
-    ],),);
-  }
-
-  Widget _buildCanvasButton({required String label, required void Function() onTap}) {
-    return GestureDetector(
-        onTap: onTap,
-        child: Padding(
-          padding: EdgeInsets.only(top: 16),
-          child: Row(
-            children: [
-              _buildDetailIcon('canvas'),
-              Expanded(
-                child: Text(label, style: Styles().textStyles.getTextStyle("widget.button.light.title.medium.underline")),
-              ),
-            ],
-          ),
-        ));
-  }
-
-  void _onLocation() {
-    Analytics().logSelect(target: "Location Directions");
-    widget.course?.launchDirections();
-  }
-
-  void _onNavigationDirections(String travelMode) {
-    Analytics().logSelect(target: 'Navigation Directions: $travelMode');
-    widget.course?.launchDirections(travelMode: travelMode);
-    //GeoMapUtils.launchDirections(destination: widget.course?.section?.building?.exploreLocation?.exploreLocationMapCoordinate, travelMode: travelMode);
-  }
-
-  void _onFloor(String floor) {
-    Analytics().logSelect(target: "Floor Plan");
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => DisplayFloorPlanPanel(building: widget.course?.section?.building, startingFloor: floor)));
-  }
-
-  void _onFloorsPlans() {
-    Analytics().logSelect(target: "Toggle Floor Plan & Amenities");
-    setState(() {
-      _floorsPlansExpanded = !_floorsPlansExpanded;
-    });
-  }
-
-  void _onCanvasAssignments() {
-    Analytics().logSelect(target: 'View Assignments in Canvas');
-    int? courseId = _canvasCourse?.id;
-    if (courseId != null) {
-      Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasCourseAssignmentsPanel(courseId: courseId, analyticsFeature: widget.analyticsFeature)));
-    }
-  }
-
-  void _onCanvasLaunch() async {
-    Analytics().logSelect(target: 'Launch Canvas');
-    String? courseDeepLinkFormat = Config().canvasCourseDeepLinkFormat;
-    String? courseDeepLink = StringUtils.isNotEmpty(courseDeepLinkFormat) ? sprintf(courseDeepLinkFormat!, [_canvasCourse?.id]) : null;
-    if (StringUtils.isNotEmpty(courseDeepLink)) {
-      await Canvas().openCanvasAppDeepLink(courseDeepLink!);
-    }
-  }
-
-  void _onMap() {
-    Analytics().logSelect(target: "View on Map");
-    NotificationService().notify(Map2.notifySelect, Map2ContentType.StudentCourses);
-  }
-
-  String? get _courseNumber => widget.course?.number;
-
-  bool get _canvasVisible => (_canvasCourse != null);
-
-  ///
-  /// #5584:
-  /// "the coursenumber property coming from the gatway building block should match the CRN of the canvas course."
-  ///
-  CanvasCourse? get _canvasCourse => (_courseNumber != null) ?
-    Canvas().courses?.firstWhereOrNull((item) => (item.crn == _courseNumber)) :
-    null;
-}
