+++ lib/ui/home/HomeStudentCoursesWidget.dart
@@ -7,8 +7,8 @@ import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/service/Auth2.dart';
 import 'package:illinois/service/Config.dart';
 import 'package:illinois/service/StudentCourses.dart';
-import 'package:illinois/ui/academics/StudentCourses.dart';
 import 'package:illinois/ui/academics/student_courses/StudentCoursesHomePanel.dart';
+import 'package:illinois/ui/academics/student_courses/StudentCoursesWidgets.dart';
 import 'package:illinois/ui/accessibility/AccessiblePageView.dart';
 import 'package:illinois/ui/home/HomePanel.dart';
 import 'package:illinois/ui/home/HomeWidgets.dart';
